Moved project into FBPainter namespace, made libpng optional, added test speeds.
All project code is declared within the FBPainter namespace to prevent name
conflicts when using this library in other projects. libpng support (and by
extension, image support) is now only enabled when USE_PNG is defined.

The test Main.cpp file now calculates how fast it moves the test image in pixels
per second. If the first argument given when launching the program is a positive
integer, it will interpret it as an alternate PPS value to use.